requirements.txt
kafka-python
python-dotenv
azure-storage-blob 
azure-identity
easydict